Move getRememberedCoupon to checkout controller
